# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Pedidos",
        "description": "API para gestionar pedidos y ventas, incluyendo la creación de pedidos, consulta de productos y entrega.",
        "contact": {
            "email": "soporte@tusistema.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "api/administradores/login": {
            "post": {
                "tags": [
                    "Administradores"
                ],
                "summary": "Iniciar sesión de administrador",
                "operationId": "af86db93a8d403fa3a3f144cab39f1fb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nickname",
                                    "password"
                                ],
                                "properties": {
                                    "nickname": {
                                        "description": "Nombre de usuario del administrador",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Contraseña del administrador",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Administrador autenticado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Administrador autenticado"
                                        },
                                        "token": {
                                            "description": "Token de acceso del administrador",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciales incorrectas"
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            }
        },
        "api/administradores/profile": {
            "get": {
                "tags": [
                    "Administradores"
                ],
                "summary": "Obtener el perfil del administrador autenticado",
                "operationId": "45a04b822b8cc63c325dae0dbf11328d",
                "responses": {
                    "200": {
                        "description": "Perfil del administrador",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Administrador"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "api/administradores/logout": {
            "post": {
                "tags": [
                    "Administradores"
                ],
                "summary": "Cerrar sesión del administrador",
                "operationId": "67f01f288615ed252bbeb62d25e3bacc",
                "responses": {
                    "200": {
                        "description": "Administrador desconectado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Administrador desconectado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Administrador": {
                "required": [
                    "id",
                    "nickname",
                    "estado",
                    "password",
                    "fecha_creacion"
                ],
                "properties": {
                    "id": {
                        "description": "ID del administrador",
                        "type": "integer"
                    },
                    "estado": {
                        "description": "Estado del administrador",
                        "type": "string",
                        "enum": [
                            "ACTIVO",
                            "INACTIVO"
                        ]
                    },
                    "fecha_creacion": {
                        "description": "Fecha de creación del administrador",
                        "type": "string",
                        "format": "date-time"
                    },
                    "fecha_actualizacion": {
                        "description": "Fecha de última actualización del administrador",
                        "type": "string",
                        "format": "date-time"
                    },
                    "nickname": {
                        "description": "Nombre de usuario del administrador",
                        "type": "string"
                    },
                    "password": {
                        "description": "Contraseña del administrador (no se debe enviar en respuestas)",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Utiliza un token de autorización Bearer para autenticar las solicitudes.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}